+++ models/vision/detection/configs/hrnet/EC2/faster_rcnn_hrnetv2p_w32_1x_coco.py
@@ -82,8 +82,8 @@ data = dict(
         pad_mode='fixed',
         preproc_mode='rgb',
         mean=(123.68, 116.78, 103.94),
-        std=(1.0, 1.0, 1.0), 
-        scale=(800, 1344)), 
+        std=(1.0, 1.0, 1.0),
+        scale=(800, 1344)),
     val=dict(
         type=dataset_type,
         train=False,
